:recycle: BottomSheet 컴포넌트를 이용해 상세페이지 BottomSheet 변경
BottomSheet 컴포넌트를 이용해 상세페이지 BottomSheet 변경
